requirements.txt
python-dotenv
requests
openai
Pillow
pydub
duckduckgo-search
